+++ internal/konnect/protocol.go
@@ -3,10 +3,8 @@ package konnect
 const (
 	// NodeTypeIngressController is the type of nodes representing KIC instances.
 	NodeTypeIngressController = "ingress-controller"
-	// NodeTypeKongProxy is the type of nodes representing normal kong gateway instances.
+	// NodeTypeKongProxy is the type of nodes representing kong gateway instances.
 	NodeTypeKongProxy = "kong-proxy"
-	// NodeTypeIngressProxy is the type of nodes representing KIC controlled kong gateway instances.
-	NodeTypeIngressProxy = "ingress-proxy"
 )
 
 type NodeItem struct {